+++ src/lib.rs
@@ -34,6 +34,7 @@
 //!   * [Fast Stochastic](indicators/struct.FastStochastic.html)
 //!   * [Slow Stochastic](indicators/struct.SlowStochastic.html)
 //!   * [Moving Average Convergence Divergence (MACD)](indicators/struct.MovingAverageConvergenceDivergence.html)
+//!   * [Percentage Price Oscillator (PPO)](indicators/struct.PercentagePriceOscillator.html)
 //!   * [Money Flow Index (MFI)](indicators/struct.MoneyFlowIndex.html)
 //! * Other
 //!   * [Standard Deviation (SD)](indicators/struct.StandardDeviation.html)