+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+#### Unreleased
+
+* Implement Percentage Price Oscillator (PPO)
+
 #### v0.2.0 - 2020-08-31
 
 * Breaking: MovingAverageConvergenceDivergence now returns MovingAverageConvergenceDivergenceOutput instead of tuple